// src/types/models.ts
export interface CoinEntry {
    BitCloutLockedNanos: number
    CoinWatermarkNanos: number;
    CoinsInCirculationNanos: number;
    CreatorBasisPoints: number;
}

export interface User {
    BalanceNanos: number;
    BlockedPubKeys: any;
    CanCreateProfile: boolean;
    HasPhoneNumber: boolean;
    IsAdmin: boolean;
    ProfileEntryResponse: Profile;
    PublicKeyBase58Check: string;
    PublicKeysBase58CheckFollowedByUser: string[];
    UsersWhoHODLYou: CreatorCoinHODLer[];
    UsersYouHODL: CreatorCoinHODLer[];
}

export interface Profile {
    ProfilePic: string;
    Username: string;
    Description: string;
    PublicKeyBase58Check: string;
    CoinPriceBitCloutNanos: number;
    CoinEntry: CoinEntry;
    IsVerified: boolean;
    CoinPriceUSD?: number;
    Posts: Post[];
    UsersThatHODL?: CreatorCoinHODLer[];
}

export interface PostReaderState {
    LikedByReader: boolean;
    RecloutedByReader: boolean;
    RecloutPostHashHex?: string;
    DiamondLevelBestowed: number;
}

export interface PostExtraData {
    EmbedVideoURL: string;
}

export interface Post {
    PostHashHex: string;
    Body: string;
    ImageURLs: string[];
    TimestampNanos: number;
    ProfileEntryResponse: Profile;
    LikeCount: number;
    PostEntryReaderState: PostReaderState;
    CommentCount: number;
    Comments: Post[];
    RecloutCount: number;
    RecloutedPostEntryResponse: Post;
    DiamondCount: number;
    ParentStakeID: string;
    PostExtraData: PostExtraData;
    ConfirmationBlockHeight: number;
    CreatorBasisPoints: number;
    InGlobalFeed: boolean;
    InMempool: boolean;
    IsHidden: boolean;
    IsPinned: boolean;
    ParentPosts: Post;
    PosterPublicKeyBase58Check: string;
    StakeMultipleBasisPoints: number;
}
// ...
export interface CreatorCoinHODLer {
    BalanceNanos: number;
    CreatorPublicKeyBase58Check: string;
    HODLerPublicKeyBase58Check: string;
    ProfileEntryResponse: Profile;
    HasPurchased: boolean;
}